Fixed: Handling of some really long qBittorrent ETAs

+++ src/NzbDrone.Core/Download/Clients/qBittorrent/QBittorrentTorrent.cs
@@ -13,7 +13,7 @@ namespace NzbDrone.Core.Download.Clients.QBittorrent
 
         public double Progress { get; set; } // Torrent progress (%/100)
 
-        public int Eta { get; set; } // Torrent ETA (seconds)
+        public ulong Eta { get; set; } // Torrent ETA (seconds)
 
         public string State { get; set; } // Torrent state. See possible values here below
